requirements.txt
requests
requests-oauthlib
twitter-text-python
